refactor(analytics): move script tag to _document.tsx
