+++ src/pages/_document.tsx
@@ -1,5 +1,6 @@
 import Document, { Html, Head, Main, NextScript, DocumentContext } from 'next/document';
 import Script from 'next/script';
+import { GOOGLE_ANALYTICS_ID } from 'lib/gtag';
 
 class MyDocument extends Document {
   static async getInitialProps(ctx: DocumentContext) {
@@ -11,6 +12,29 @@ class MyDocument extends Document {
     return (
       <Html lang="ja">
         <Head>
+          {GOOGLE_ANALYTICS_ID && (
+            <>
+              <Script
+                id="gtag-manager"
+                strategy="afterInteractive"
+                src={`https://www.googletagmanager.com/gtag/js?id=${GOOGLE_ANALYTICS_ID}`}
+              />
+              <Script
+                id="gtag-script"
+                strategy="afterInteractive"
+                dangerouslySetInnerHTML={{
+                  __html: `
+                    window.dataLayer = window.dataLayer || [];
+                    function gtag(){dataLayer.push(arguments);}
+                    gtag('js', new Date());
+                    gtag('config', '${GOOGLE_ANALYTICS_ID}', {
+                      page_path: window.location.pathname,
+                    });
+                  `,
+                }}
+              />
+            </>
+          )}
           <Script
             data-ad-client={`${process.env.GOOGLE_ADSENSE_CLIENT_ID}`}
             async
